+++ src/main/java/org/jabref/gui/openoffice/OpenOfficePanel.java
@@ -8,6 +8,7 @@ import java.util.Optional;
 
 import javax.swing.undo.UndoManager;
 
+import javafx.beans.property.SimpleObjectProperty;
 import javafx.concurrent.Task;
 import javafx.geometry.Insets;
 import javafx.geometry.Side;
@@ -64,6 +65,7 @@ import org.jabref.model.util.FileUpdateMonitor;
 import com.sun.star.comp.helper.BootstrapException;
 import com.sun.star.container.NoSuchElementException;
 import com.sun.star.lang.WrappedTargetException;
+import com.tobiasdiez.easybind.EasyBind;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -101,8 +103,11 @@ public class OpenOfficePanel {
     private final LibraryTabContainer tabContainer;
     private final FileUpdateMonitor fileUpdateMonitor;
     private final BibEntryTypesManager entryTypesManager;
+    private final OpenOfficePreferences openOfficePreferences;
     private OOBibBase ooBase;
     private OOStyle currentStyle;
+
+    private final SimpleObjectProperty<OOStyle> currentStyleProperty;
 
     public OpenOfficePanel(LibraryTabContainer tabContainer,
                            GuiPreferences preferences,
@@ -124,6 +129,7 @@ public class OpenOfficePanel {
         this.clipBoardManager = clipBoardManager;
         this.undoManager = undoManager;
         this.currentStyle = preferences.getOpenOfficePreferences().getCurrentStyle();
+        this.openOfficePreferences = preferences.getOpenOfficePreferences();
 
         ActionFactory factory = new ActionFactory();
 
@@ -155,6 +161,8 @@ public class OpenOfficePanel {
                 preferences.getLayoutFormatterPreferences(),
                 abbreviationRepository);
 
+        currentStyleProperty = new SimpleObjectProperty<>(currentStyle);
+
         initPanel();
     }
 
@@ -168,6 +176,7 @@ public class OpenOfficePanel {
      */
     private boolean getOrUpdateTheStyle(String title) {
         currentStyle = loader.getUsedStyleUnified();
+        currentStyleProperty.set(currentStyle);
         final boolean FAIL = true;
         final boolean PASS = false;
 
@@ -213,6 +222,8 @@ public class OpenOfficePanel {
             dialogService.showCustomDialogAndWait(styleDialog)
                          .ifPresent(selectedStyle -> {
                              currentStyle = selectedStyle;
+                             currentStyleProperty.set(currentStyle);
+
                              if (currentStyle instanceof JStyle jStyle) {
                                  try {
                                      jStyle.ensureUpToDate();
@@ -470,7 +481,7 @@ public class OpenOfficePanel {
     }
 
     private OOBibBase createBibBase(Path loPath) throws BootstrapException, CreationException {
-        return new OOBibBase(loPath, dialogService);
+        return new OOBibBase(loPath, dialogService, openOfficePreferences);
     }
 
     /**
@@ -612,12 +623,27 @@ public class OpenOfficePanel {
     }
 
     private ContextMenu createSettingsPopup() {
-        OpenOfficePreferences openOfficePreferences = preferences.getOpenOfficePreferences();
-
         ContextMenu contextMenu = new ContextMenu();
 
         CheckMenuItem autoSync = new CheckMenuItem(Localization.lang("Automatically sync bibliography when inserting citations"));
-        autoSync.selectedProperty().set(preferences.getOpenOfficePreferences().getSyncWhenCiting());
+        autoSync.selectedProperty().set(openOfficePreferences.getSyncWhenCiting());
+
+        CheckMenuItem alwaysAddCitedOnPagesText = new CheckMenuItem(Localization.lang("Automatically add \"Cited on pages...\" at the end of bibliographic entries"));
+        alwaysAddCitedOnPagesText.selectedProperty().set(openOfficePreferences.getAlwaysAddCitedOnPages());
+        alwaysAddCitedOnPagesText.setOnAction(e -> openOfficePreferences.setAlwaysAddCitedOnPages(alwaysAddCitedOnPagesText.isSelected()));
+
+        EasyBind.listen(currentStyleProperty, (obs, oldValue, newValue) -> {
+            switch (newValue) {
+                case JStyle ignored -> {
+                    if (!contextMenu.getItems().contains(alwaysAddCitedOnPagesText)) {
+                        contextMenu.getItems().add(1, alwaysAddCitedOnPagesText);
+                    }
+                }
+                case CitationStyle ignored ->
+                        contextMenu.getItems().remove(alwaysAddCitedOnPagesText);
+                default -> { }
+            }
+        });
 
         ToggleGroup toggleGroup = new ToggleGroup();
         RadioMenuItem useActiveBase = new RadioMenuItem(Localization.lang("Look up BibTeX entries in the active tab only"));
@@ -648,6 +674,10 @@ public class OpenOfficePanel {
                 useAllBases,
                 new SeparatorMenuItem(),
                 clearConnectionSettings);
+
+        if (currentStyle instanceof JStyle) {
+            contextMenu.getItems().add(1, alwaysAddCitedOnPagesText);
+        }
 
         return contextMenu;
     }